extract a util to generate staging path and update arch violations
