+++ flink-connectors/flink-connector-files/src/test/java/org/apache/flink/connector/file/table/FileSystemOutputFormatTest.java
@@ -22,7 +22,6 @@ import org.apache.flink.api.common.io.FinalizeOnMaster.FinalizationContext;
 import org.apache.flink.api.java.io.TextOutputFormat;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.streaming.api.functions.sink.OutputFormatSinkFunction;
-import org.apache.flink.streaming.api.functions.sink.filesystem.OutputFileConfig;
 import org.apache.flink.streaming.api.operators.StreamSink;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 import org.apache.flink.streaming.util.OneInputStreamOperatorTestHarness;
@@ -34,24 +33,16 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
-import java.util.UUID;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
@@ -113,101 +104,15 @@ class FileSystemOutputFormatTest {
                 .containsExactly("a1,1,p1\n" + "a2,2,p1\n" + "a2,2,p2\n" + "a3,3,p1\n");
     }
 
-    @ParameterizedTest
-    @ValueSource(booleans = {true, false})
-    void testStagingDirBehavior(boolean shareStagingDir) throws Exception {
-        // sink1
-        AtomicReference<FileSystemOutputFormat<Row>> sinkRef1 = new AtomicReference<>();
-        AtomicReference<Map<File, String>> fileToCommitRef1 = new AtomicReference<>();
-        writeRowsToSink(
-                sinkRef1,
-                fileToCommitRef1,
-                getStagingDir(shareStagingDir),
-                Row.of("a1", 1, "x1"),
-                Row.of("a2", 2, "x2"));
-
-        // sink2
-        AtomicReference<FileSystemOutputFormat<Row>> sinkRef2 = new AtomicReference<>();
-        AtomicReference<Map<File, String>> fileToCommitRef2 = new AtomicReference<>();
-        writeRowsToSink(
-                sinkRef2,
-                fileToCommitRef2,
-                getStagingDir(shareStagingDir),
-                Row.of("b1", 1, "y1"),
-                Row.of("b2", 2, "y2"));
-
-        assertSinkBehavior(sinkRef1, fileToCommitRef1, sinkRef2, fileToCommitRef2, shareStagingDir);
-    }
-
-    private void writeRowsToSink(
-            AtomicReference<FileSystemOutputFormat<Row>> sinkRef,
-            AtomicReference<Map<File, String>> contentRef,
-            Path stagingDir,
-            Row... records)
-            throws Exception {
-        try (OneInputStreamOperatorTestHarness<Row, Object> testHarness =
-                createSink(false, false, false, stagingDir, new LinkedHashMap<>(), sinkRef)) {
-            writeUnorderedRecords(testHarness, Arrays.asList(records));
-            contentRef.set(getFileContentByPath(Paths.get(stagingDir.getPath())));
-        }
-    }
-
-    private Path getStagingDir(boolean shareStagingDir) {
-        String pathPostfix = FileSystemTableSink.getStagingPathPostfix(shareStagingDir);
-        return Path.fromLocalFile(tmpPath.resolve(pathPostfix).toFile());
-    }
-
-    private void assertSinkBehavior(
-            AtomicReference<FileSystemOutputFormat<Row>> sinkRef1,
-            AtomicReference<Map<File, String>> fileToCommitRef1,
-            AtomicReference<FileSystemOutputFormat<Row>> sinkRef2,
-            AtomicReference<Map<File, String>> fileToCommitRef2,
-            boolean shareStagingDir)
-            throws Exception {
-        Map<File, String> fileToCommit1 = fileToCommitRef1.get();
-        Map<File, String> fileToCommit2 = fileToCommitRef2.get();
-        assertThat(fileToCommit2.keySet()).allMatch(File::exists);
-        if (shareStagingDir) {
-            assertThat(fileToCommit1.keySet()).noneMatch(File::exists);
-        } else {
-            assertThat(fileToCommit1.keySet()).allMatch(File::exists);
-        }
-        sinkRef1.get().finalizeGlobal(finalizationContext);
-        sinkRef2.get().finalizeGlobal(finalizationContext);
-        Map<File, String> committedFiles = getFileContentByPath(outputPath);
-        if (shareStagingDir) {
-            assertThat(committedFiles).hasSize(1);
-            assertThat(committedFiles.values()).containsExactlyElementsOf(fileToCommit2.values());
-        } else {
-            assertThat(committedFiles).hasSize(2);
-            assertThat(committedFiles.values())
-                    .containsExactlyInAnyOrderElementsOf(
-                            Stream.concat(
-                                            fileToCommit1.values().stream(),
-                                            fileToCommit2.values().stream())
-                                    .collect(Collectors.toList()));
-        }
-    }
-
     private void writeUnorderedRecords(OneInputStreamOperatorTestHarness<Row, Object> testHarness)
-            throws Exception {
-        List<Row> records =
-                Arrays.asList(
-                        Row.of("a1", 1, "p1"),
-                        Row.of("a2", 2, "p1"),
-                        Row.of("a2", 2, "p2"),
-                        Row.of("a3", 3, "p1"));
-        writeUnorderedRecords(testHarness, records);
-    }
-
-    private void writeUnorderedRecords(
-            OneInputStreamOperatorTestHarness<Row, Object> testHarness, List<Row> records)
             throws Exception {
         testHarness.setup();
         testHarness.open();
-        for (Row record : records) {
-            testHarness.processElement(new StreamRecord<>(record, 1L));
-        }
+
+        testHarness.processElement(new StreamRecord<>(Row.of("a1", 1, "p1"), 1L));
+        testHarness.processElement(new StreamRecord<>(Row.of("a2", 2, "p1"), 1L));
+        testHarness.processElement(new StreamRecord<>(Row.of("a2", 2, "p2"), 1L));
+        testHarness.processElement(new StreamRecord<>(Row.of("a3", 3, "p1"), 1L));
     }
 
     @Test
@@ -304,7 +209,6 @@ class FileSystemOutputFormatTest {
             boolean override,
             boolean partition,
             boolean dynamicGrouped,
-            Path stagingDir,
             LinkedHashMap<String, String> staticPartitions,
             AtomicReference<FileSystemOutputFormat<Row>> sinkRef)
             throws Exception {
@@ -316,11 +220,7 @@ class FileSystemOutputFormatTest {
         FileSystemOutputFormat<Row> sink =
                 new FileSystemOutputFormat.Builder<Row>()
                         .setMetaStoreFactory(msFactory)
-                        .setOutputFileConfig(
-                                OutputFileConfig.builder()
-                                        .withPartPrefix("part-" + UUID.randomUUID())
-                                        .build())
-                        .setTempPath(stagingDir)
+                        .setTempPath(new Path(tmpPath.toString()))
                         .setOverwrite(override)
                         .setPartitionColumns(partitionColumns)
                         .setPartitionComputer(
@@ -338,22 +238,6 @@ class FileSystemOutputFormatTest {
                 3,
                 3,
                 0);
-    }
-
-    private OneInputStreamOperatorTestHarness<Row, Object> createSink(
-            boolean override,
-            boolean partition,
-            boolean dynamicGrouped,
-            LinkedHashMap<String, String> staticPartitions,
-            AtomicReference<FileSystemOutputFormat<Row>> sinkRef)
-            throws Exception {
-        return createSink(
-                override,
-                partition,
-                dynamicGrouped,
-                new Path(tmpPath.toString()),
-                staticPartitions,
-                sinkRef);
     }
 
     private static class TestingFinalizationContext implements FinalizationContext {
